feat(shell-tools): F4.6.7 phase 1 — 提取 shell.rs 沙箱无关辅助到 dasclaw_shell_tools
ADR-156 §6.3 F4.6.7 phase 1。把 desktop-client/ironclaw/src/tools/builtin/shell.rs
（1666 行）中和沙箱无关的纯函数 / 静态量 verbatim 抽到新 crate dasclaw_shell_tools，
desktop 通过 `use dasclaw_shell_tools::...` 引用，函数体未改一字。

迁出的 API：
- 静态量：MAX_OUTPUT_SIZE / BLOCKED_COMMANDS / DANGEROUS_PATTERNS / SAFE_ENV_VARS
- 风险分类：classify_command_risk / analyze_command_for_result
- 注入检测：detect_command_injection
- 输出处理：truncate_output / truncate_for_error
- 参数提取：extract_command_param

对应单测（23 条）整体迁过来。desktop 侧保留 ShellTool 集成测试（32 条）。

非目标（下个 phase 处理）：
- 不动 ShellTool struct 本体
- 不动 OsExecutor / desktop 局部 SandboxPolicy enum

Cross-cuts: 类A（无新 .ironclaw 字面量）

验证：
- cargo check -p dasclaw_shell_tools --tests ✅
- cargo check -p desktop-client --tests ✅
- cargo nextest run -p dasclaw_shell_tools ✅ 23/23
- cargo nextest run -p dasclaw -E 'test(/shell/)' ✅ 32/32
- cargo fmt --all ✅
- python3.12 scripts/check_no_panics.py --base origin/xClaw ✅
- cargo clippy --no-deps -p dasclaw_shell_tools --all-targets -- -D warnings ✅
- cargo clippy --no-deps -p dasclaw --all-targets -- -D warnings ✅

Closes #783
