+++ Cargo.toml
@@ -45,6 +45,8 @@ members = [
 	"crates/dasclaw_fs_tools",
 	# F4.6.4 — sub_agent builtin tool (SubAgentTool / SubAgentRole) per ADR-156 §6.3
 	"crates/dasclaw_sub_agent_tools",
+	# F4.6.7 Phase 1 — shell-tool sandbox-independent helpers per ADR-156 §6.3
+	"crates/dasclaw_shell_tools",
 	# W2.6 接线层：组合 sandbox + pty + workspace_cap policy
 	"crates/dasclaw_exec",
 	# W6-C 主仓自实现 LLM provider —— 替代 claw-code-api path-dep（ADR-118 PR-A.0 骨架）
